Vectors in float

+++ game00.py
@@ -19,11 +19,11 @@ YELLOW = (255, 255, 0)
 
 @dataclass
 class Player():
-    x: int
-    y: int
+    x: float
+    y: float
 
     direction: str = ""
-    speed: int = 4
+    speed: float = 4
     alive: bool = True
     score: int = 0
