+++ game01.py
@@ -19,11 +19,11 @@ YELLOW = (255, 255, 0)
 
 @dataclass
 class Player():
-    x: int
-    y: int
+    x: float
+    y: float
 
     direction: str = ""
-    speed: int = 4
+    speed: float = 4
     alive: bool = True
     score: int = 0
 
@@ -45,10 +45,10 @@ class Player():
 
 @dataclass()
 class Bullet():
-    x: int
-    y: int
+    x: float
+    y: float
     is_shot: bool = False
-    speed: int = 4
+    speed: float = 4
 
     def shoot(self):
         # 2.1 Action a faire lorsque le joueur tire
